Fix parsing of bundle cli json output
cfssljson was not able to parse the output of the cfssl bundle command.
The existing bundle parsing expected the structure to be
`input["result"]["bundle"]["bundle"]` but the output from the cli is
simply `input["bundle"]`.  The updated parsing still supports the old
structure in case this is still being used elsewhere.